The Use of Twitter API for Searching of Keywords
This repository consists of the Jupyter notebook file for doing the analysis of the twitter trends. You can easily check this code out and try on your own for looking for twitter trends. We used the Twitter API for the same. You just have to make an app in twitter. from where you will get the credentials for the same. And which you have to write in the places as in the notebook. 
You can fetch the keywords from the twitter and find the count of the words which are used most in those tweets. And much more you can do with this code.